backend/requirements.txt
alabaster==0.7.12
alembic==1.8.1
aniso8601==6.0.0
apturl==0.5.2
attrs==19.3.0
Babel==2.9.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
Camelot==12.6.29
camelot-py==0.8.2
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
cliapp==1.20180812.1
click==8.1.3
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distlib==0.3.5
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
docutils==0.17.1
duplicity==0.8.12.0
Elixir==0.7.1
entrypoints==0.3
et-xmlfile==1.1.0
fasteners==0.14.1
filelock==3.8.0
flacon==0.5.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
fonttools==4.28.5
future==0.18.2
GDAL==3.0.4
gpg===1.13.1-unknown
greenlet==1.1.2
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
ifaddr==0.1.6
imagesize==1.2.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
install==1.3.4
ipython_genutils==0.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
jupyter-core==4.6.3
kdtree==0.16
keyring==18.0.1
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.1.1
matplotlib==3.5.1
monotonic==1.5
more-itertools==4.2.0
nbformat==5.0.4
netifaces==0.10.4
nose==1.3.7
numpy==1.22.1
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.7
openshot-qt==2.4.3
OWSLib==0.19.1
packaging==20.3
pandas==1.2.4
paramiko==2.6.0
pb-tool==3.1.0
pbr==5.6.0
pdfarranger==1.4.2
pdfminer.six==20201018
pdftables-api==1.1.0
pexpect==4.6.0
pikepdf==1.10.3+dfsg
Pillow==8.3.1
platformdirs==2.5.2
plotly==4.4.1
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.3
protobuf==3.6.1
psutil==5.5.1
psycopg2-binary==2.9.3
pycairo==1.16.2
PyChromecast==4.1.0
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
PyPDF2==1.26.0
pyproj==2.5.0
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.6.0
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
pytz==2019.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
raven==6.10.0
requests==2.22.0
requests-unixsocket==0.2.0
retrying==1.3.3
rpy2==3.4.5
scipy==1.7.1
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.12.0
snowballstemmer==2.1.0
sortedcontainers==2.4.0
soupsieve==1.9.5
Sphinx==4.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.3.4
sqlalchemy-migrate==0.13.0
sqlparse==0.4.1
ssh-import-id==5.10
statistic==0.1.6
system-service==0.3
systemd-python==234
tabula==1.0.5
tabula-py==2.2.0
Tempita==0.5.2
traitlets==4.3.3
ttystatus==0.38
tzlocal==3.0
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.16.3
wadllib==1.3.3
webencodings==0.5.1
Werkzeug==2.2.2
WTForms==3.0.1
xkit==0.0.0
xlrd==0.7.1
xlwt==0.7.2
youtube-dl==2020.3.24
zeroconf==0.24.4
zipp==3.8.1
